v1.0

+++ app/Controllers/Http/ClassroomsController.ts
@@ -1,6 +1,7 @@
 import type { HttpContextContract } from "@ioc:Adonis/Core/HttpContext";
 import Classroom from "App/Models/Classroom";
-import AddStudentValidator from "App/Validators/ClassRooms/AddStudentValidator";
+import AttachStudentValidator from "App/Validators/ClassRooms/AttachStudentValidator";
+import AddStudentValidator from "App/Validators/ClassRooms/AttachStudentValidator";
 import CreateClassroomValidator from "App/Validators/ClassRooms/CreatsClassroomValidator";
 import UpdateClassRoomValidator from "App/Validators/ClassRooms/UpdateClassRoomValidator";
 
@@ -8,7 +9,13 @@ export default class ClassroomsController {
   public async create(ctx: HttpContextContract) {
     const payload = await ctx.request.validate(CreateClassroomValidator);
 
-    const classroom = await Classroom.create(payload);
+    const user = ctx.auth.user!;
+
+    const classroom = await Classroom.create({
+      number: payload.number,
+      capacity: payload.capacity,
+      teacherId: user.id,
+    });
 
     return ctx.response.created(classroom);
   }
@@ -16,9 +23,16 @@ export default class ClassroomsController {
   public async update(ctx: HttpContextContract) {
     const { id } = ctx.params;
 
-    const payload = await ctx.request.validate(UpdateClassRoomValidator);
-
+    const user = ctx.auth.user!;
     const classroom = await Classroom.query().where("id", id).firstOrFail();
+
+    if (user.id !== classroom.teacherId) {
+      return ctx.response.badRequest(
+        "O professor não é o administrador da sala"
+      );
+    }
+
+    const payload = await ctx.request.validate(UpdateClassRoomValidator);
 
     classroom.merge(payload).save();
 
@@ -28,17 +42,23 @@ export default class ClassroomsController {
   public async findOne(ctx: HttpContextContract) {
     const { id } = ctx.params;
 
+    const user = ctx.auth.user!;
+
     const classroom = await Classroom.query()
       .where("id", id)
-      .preload("teacher")
-      .preload("students")
+      .preload("teacher", (query) => {
+        query.select("id", "name", "email", "registration");
+      })
+      .preload("students", (query) => {
+        query.select("id", "name", "email", "registration");
+      })
       .firstOrFail();
 
-    return ctx.response.ok(classroom);
-  }
-
-  public async findAll(ctx: HttpContextContract) {
-    const classroom = await Classroom.query();
+    if (user.id !== classroom.teacherId) {
+      return ctx.response.badRequest(
+        "O professor não é o administrador da sala"
+      );
+    }
 
     return ctx.response.ok(classroom);
   }
@@ -46,20 +66,74 @@ export default class ClassroomsController {
   public async delete(ctx: HttpContextContract) {
     const { id } = ctx.params;
 
-    const classroom = await Classroom.findOrFail(id);
+    const user = ctx.auth.user!;
+    const classroom = await Classroom.query().where("id", id).firstOrFail();
+
+    if (user.id !== classroom.teacherId) {
+      return ctx.response.badRequest(
+        "O professor não é o administrador da sala"
+      );
+    }
 
     classroom.delete();
 
     return ctx.response.noContent();
   }
-  public async addOne(ctx: HttpContextContract) {
-    const payload = await ctx.request.validate(AddStudentValidator);
+
+  public async attachStudent(ctx: HttpContextContract) {
+    const payload = await ctx.request.validate(AttachStudentValidator);
+
+    const user = ctx.auth.user!;
 
     const classroom = await Classroom.query()
       .where("id", payload.classroomId)
       .firstOrFail();
 
+    if (user.id !== classroom.teacherId) {
+      return ctx.response.badRequest(
+        "O professor não é o administrador da sala"
+      );
+    }
+
+    const studentExistInClassroom = await Classroom.query()
+      .where("id", payload.classroomId)
+      .andWhereHas("students", (query) => query.where("id", payload.studentId))
+      .first();
+
+    if (studentExistInClassroom) {
+      return ctx.response.badRequest("O Aluno já está matriculado nessa sala");
+    }
+
     classroom.related("students").attach([payload.studentId]);
+
+    return ctx.response.ok(classroom);
+  }
+
+  public async detachStudent(ctx: HttpContextContract) {
+    const payload = await ctx.request.validate(AttachStudentValidator);
+
+    const user = ctx.auth.user!;
+
+    const classroom = await Classroom.query()
+      .where("id", payload.classroomId)
+      .firstOrFail();
+
+    if (user.id !== classroom.teacherId) {
+      return ctx.response.badRequest(
+        "O professor não é o administrador da sala"
+      );
+    }
+
+    const studentExistInClassroom = await Classroom.query()
+      .where("id", payload.classroomId)
+      .andWhereHas("students", (query) => query.where("id", payload.studentId))
+      .first();
+
+    if (!studentExistInClassroom) {
+      return ctx.response.badRequest("O Aluno não está matriculado nessa sala");
+    }
+
+    classroom.related("students").detach([payload.studentId]);
 
     return ctx.response.ok(classroom);
   }
